add links to external sources
Adding links to referenced texts gives readers an easy way to go to a referenced document to learn more about it.

For example, when the document says:

> "In General.—Subtitle VI of title 31, United States Code, is amended by adding after chapter 97 the following:"

It's helpful to provide the reader with a  link to [Subtitle VI of title 31](https://www.law.cornell.edu/uscode/text/31/subtitle-VI) and [chapter 97](https://www.law.cornell.edu/uscode/text/31/subtitle-VI/chapter-97).

Or when the document says:
> "(7) Security.—The term ‘security’ has the meaning given the term in section 3(a) of the Securities Exchange Act of 1934 (15 U.S.C. 78c(a)).

It's helpful to link to [section 3(a) of the Securities Exchange Act of 1934](https://www.govinfo.gov/content/pkg/COMPS-1885/pdf/COMPS-1885.pdf#page=11) and even better when we can link to the specific page.

In general, I tried to pick `.gov` sources since it would be the most authoritative source, but in some cases, `law.cornell.edu` was a better reference. 

I only had enough time to add a couple links but I want to use this pull request as a starting point to inspire others to add more links.